Move SubmitRequestAction and DeleteRequestAction to specimen-main
