+++ api/src/org/labkey/api/study/SpecimenUrls.java
@@ -30,7 +30,6 @@ import org.springframework.web.servlet.mvc.Controller;
 public interface SpecimenUrls extends UrlProvider
 {
     ActionURL getCommentURL(Container c, String globalUniqueId);
-    ActionURL getDeleteRequestURL(Container c, String id);
     ActionURL getInsertSpecimenQueryRowURL(Container c, String schemaName, TableInfo table);
     ActionURL getManageRequestURL(Container c, int requestId);
     @Migrate // Eliminate this -- same as getManageRequestURL(Container, int);
@@ -38,8 +37,6 @@ public interface SpecimenUrls extends UrlProvider
     ActionURL getManageRequestStatusURL(Container c, int requestId);
     ActionURL getSpecimensURL(Container c);
     ActionURL getSpecimensURL(Container c, boolean showVials);
-    ActionURL getSelectedSpecimensURL(Container c, boolean showVials);
-    ActionURL getSubmitRequestURL(Container c, String id);
     ActionURL getUpdateSpecimenQueryRowURL(Container c, String schemaName, TableInfo table);
 
     Class<? extends Controller> getCopyParticipantCommentActionClass();
