+++ specimen/src/org/labkey/specimen/actions/SpecimenController2.java
@@ -58,6 +58,7 @@ import org.labkey.api.specimen.RequestEventType;
 import org.labkey.api.specimen.RequestedSpecimens;
 import org.labkey.api.specimen.SpecimenManager;
 import org.labkey.api.specimen.SpecimenManagerNew;
+import org.labkey.api.specimen.SpecimenMigrationService;
 import org.labkey.api.specimen.SpecimenRequestException;
 import org.labkey.api.specimen.SpecimenRequestManager;
 import org.labkey.api.specimen.SpecimenRequestStatus;
@@ -141,6 +142,7 @@ import org.labkey.specimen.security.permissions.ManageRequestRequirementsPermiss
 import org.labkey.specimen.view.SpecimenSearchWebPart;
 import org.springframework.validation.BindException;
 import org.springframework.validation.Errors;
+import org.springframework.validation.ObjectError;
 import org.springframework.web.servlet.ModelAndView;
 
 import javax.mail.Address;
@@ -3746,6 +3748,106 @@ public class SpecimenController2 extends SpringActionController
         {
             addBaseSpecimenNavTrail(root);
             root.addChild("Set vial comments");
+        }
+    }
+
+    @RequiresPermission(RequestSpecimensPermission.class)
+    public class SubmitRequestAction extends FormHandlerAction<IdForm>
+    {
+        @Override
+        public void validateCommand(IdForm target, Errors errors)
+        {
+        }
+
+        @Override
+        public boolean handlePost(IdForm form, BindException errors) throws Exception
+        {
+            if (!SettingsManager.get().isSpecimenShoppingCartEnabled(getContainer()))
+                throw new UnauthorizedException();
+
+            SpecimenRequest request = SpecimenRequestManager.get().getRequest(getContainer(), form.getId());
+            requiresEditRequestPermissions(request);
+            List<Vial> vials = request.getVials();
+            if (!vials.isEmpty() || SpecimenService.get().getRequestCustomizer().allowEmptyRequests())
+            {
+                SpecimenRequestStatus newStatus = SpecimenRequestManager.get().getInitialRequestStatus(getContainer(), getUser(), true);
+                request = request.createMutable();
+                request.setStatusId(newStatus.getRowId());
+                try
+                {
+                    SpecimenRequestManager.get().updateRequest(getUser(), request);
+                    SpecimenRequestManager.get().createRequestEvent(getUser(), request,
+                            RequestEventType.REQUEST_STATUS_CHANGED, "Request submitted for processing.", null);
+                    if (SettingsManager.get().isSpecimenShoppingCartEnabled(getContainer()))
+                    {
+                        try
+                        {
+                            StudyInternalService.get().sendNewRequestNotifications(getViewContext(), request, errors);
+                        }
+                        catch (ConfigurationException e)
+                        {
+                            errors.reject(ERROR_MSG, e.getMessage());
+                        }
+                    }
+                }
+                catch (RequestabilityManager.InvalidRuleException e)
+                {
+                    errors.reject(ERROR_MSG, "The request could not be submitted because a requestability rule is configured incorrectly. " +
+                            "Please report this problem to an administrator. Error details: "  + e.getMessage());
+                }
+                catch (IOException e)
+                {
+                    errors.reject(ERROR_MSG, e.getMessage());
+                }
+            }
+            else
+            {
+                errors.addError(new ObjectError("Specimen Request", new String[] {"NullError"}, null, "Only requests containing specimens can be submitted."));
+            }
+
+            return !errors.hasErrors();
+        }
+
+        @Override
+        public ActionURL getSuccessURL(IdForm idForm)
+        {
+            return SpecimenMigrationService.get().getManageRequestURL(getContainer(), idForm.getId(), null).addParameter("submissionResult", true);
+        }
+    }
+
+    @RequiresPermission(RequestSpecimensPermission.class)
+    public class DeleteRequestAction extends FormHandlerAction<IdForm>
+    {
+        @Override
+        public void validateCommand(IdForm target, Errors errors)
+        {
+        }
+
+        @Override
+        public boolean handlePost(IdForm form, BindException errors) throws Exception
+        {
+            SpecimenRequest request = SpecimenRequestManager.get().getRequest(getContainer(), form.getId());
+            requiresEditRequestPermissions(request);
+            SpecimenRequestStatus cartStatus = SpecimenRequestManager.get().getRequestShoppingCartStatus(getContainer(), getUser());
+            if (request.getStatusId() != cartStatus.getRowId())
+                throw new UnauthorizedException();
+
+            try
+            {
+                SpecimenRequestManager.get().deleteRequest(getUser(), request);
+            }
+            catch (RequestabilityManager.InvalidRuleException e)
+            {
+                errors.reject(ERROR_MSG, "The request could not be canceled because a requestability rule is configured incorrectly. " +
+                        "Please report this problem to an administrator. Error details: "  + e.getMessage());
+            }
+            return true;
+        }
+
+        @Override
+        public ActionURL getSuccessURL(IdForm form)
+        {
+            return new ActionURL(ViewRequestsAction.class, getContainer());
         }
     }
 }
